Update to match SuperCORE
