--- vendordeps/NavX.json
@@ -1,39 +1,0 @@
-{
-    "fileName": "NavX.json",
-    "name": "KauaiLabs_navX_FRC",
-    "version": "2023.0.3",
-    "uuid": "cb311d09-36e9-4143-a032-55bb2b94443b",
-    "mavenUrls": [
-        "https://dev.studica.com/maven/release/2023/"
-    ],
-    "jsonUrl": "https://dev.studica.com/releases/2023/NavX.json",
-    "javaDependencies": [
-        {
-            "groupId": "com.kauailabs.navx.frc",
-            "artifactId": "navx-frc-java",
-            "version": "2023.0.3"
-        }
-    ],
-    "jniDependencies": [],
-    "cppDependencies": [
-        {
-            "groupId": "com.kauailabs.navx.frc",
-            "artifactId": "navx-frc-cpp",
-            "version": "2023.0.3",
-            "headerClassifier": "headers",
-            "sourcesClassifier": "sources",
-            "sharedLibrary": false,
-            "libName": "navx_frc",
-            "skipInvalidPlatforms": true,
-            "binaryPlatforms": [
-                "linuxathena",
-                "linuxraspbian",
-                "linuxarm32",
-                "linuxarm64",
-                "linuxx86-64",
-                "osxuniversal",
-                "windowsx86-64"
-            ]
-        }
-    ]
-}
